feat: security hardening, ClamAV deep integration, additional-files reorganization
- ClamAV fail-open with CriticalSystemError event/listener pipeline
- Email + Slack admin notifications with queued retry backoff
- Dismissible frontend maintenance popup on system degradation
- Organized 120+ supplementary files into additional-files/ structure
- Fixed AuditLog target_user_id constraint and sockets extension guard
- Updated .cursorrules with additional-files placement policy
- PHPUnit tests passing for malware block and fail-open scenarios

+++ database/migrations/2026_03_26_121504_create_elite_subscriptions_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('elite_subscriptions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('shop_id');
+            $table->integer('shop_id');
             $table->decimal('amount_paid', 10, 2)->default(0);
             $table->enum('billing_cycle', ['monthly', 'yearly'])->default('monthly');
             $table->enum('status', ['pending', 'active', 'expired', 'rejected'])->default('pending');
